+++ backend/image_converter/core/enums/image_format.py
@@ -1,10 +1,11 @@
 from enum import Enum
-from backend.image_converter.core.internals.utls import Result
+from backend.image_converter.core.internals.utilities import Result
 
 class ImageFormat(Enum):
     JPEG = "JPEG"
     PNG = "PNG"
     ICO = "ICO"
+    AVIF = "AVIF"
 
     @classmethod
     def from_string(cls, value: str) -> "ImageFormat":
@@ -40,4 +41,5 @@ IMAGE_FORMAT_EXTENSIONS = {
     "JPEG": ".jpg",
     "PNG": ".png",
     "ICO": ".ico",
+    "AVIF": ".avif",
 }