+++ backend/image_converter/infrastructure/cleanup_service.py
@@ -3,7 +3,7 @@ import time
 import shutil
 from typing import Dict, List
 import traceback
-from backend.image_converter.core.internals.utls import Result                                       
+from backend.image_converter.core.internals.utilities import Result                                       
 
 class CleanupService:
     """